requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
pydantic==2.11.10
pydantic-settings==2.11.0
email-validator==2.2.0
anthropic==0.69.0
mcp==1.16.0
httpx==0.28.1
python-dotenv==1.0.1